Fix cast.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@22705 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/kernel/arch/m68k/arch_thread.h
@@ -25,7 +25,7 @@ arch_thread_get_current_thread(void)
 {
 	uint64 v = 0;
 	asm volatile("pmove %%srp,(%0)" : : "a"(&v));
-	return (struct thread *)(v & 0xffffffff);
+	return (struct thread *)(uint32)(v & 0xffffffff);
 }
